Implement embedded_hal::Pwm for TIMER0..=5
This is modeled by the `serial` module, and takes ownership of timer
peripherals and the appropriate GPIO pins to implement the Pwm trait.
It is slightly unfortunate that it takes ownership of the entire timer
peripheral rather than allowing the user to call `.into_even()` and use
timer B for other things (or vice-versa).

This has been tested by porting the examples in the stellaris-launchpad
crate and running them on the EK-LM4F120XL LaunchPad.